Tests: add NUnit support
What's an open source project that cannot be
tested in an open source platform? The best
open source .NET IDE that works in Linux
(MonoDevelop) doesn't have support for
VisualStudio Testing framework, but it has
support for NUnit.

So, as it was before, MonoDevelop users could
not run the tests.

Now the Tests assembly references NUnit, and
a new TLSharp.Tests.VS assembly is added which
wraps the NUnit ones, so that VS users that
don't want to install a NUnit plugin can still
run them.

By having NUnit referenced in the main Tests
assembly, Visual Studio can still open the
project, but the other way around is not
possible (MonoDevelop users cannot open a
MSTest assembly). So this solution is the best
of both worlds.