Added two methods - GetUserProfileImage and GetUserPhotos
Had to write this for my project and now i suggest you adding them here. You may want to check catch blocks in my methods to ensure they are correct. There is quite a mess in GetUserProfileImage method but it seems to work properly, so that i decided to suggest it not matter what.