Core: save session as soon as a the sequence is incremented
Just be safe, let's not wait until other part of the code saves the
session, and let's do it right here on the spot, where we know it
changes.

@Laituex was having a ErrCode=32 'msg_seqno too low' problem and this
is the second thing we suspected to be the culprit, because if there's
any issue (e.g. exception) that makes the session not be saved after
the sequence got incremented, then an old sequence would be used the
next time upon loading an old session file.

(Backported from https://github.com/sochix/TLSharp/commit/1b06473108fb41e1b1fc5913b63782edf240a495)